get activities and online moved to external files
