+++ clan/clanMember/clanMember.js
@@ -213,63 +213,3 @@ export function GetDiscordMemberByMention(guild, discordMention) {
     console.log(discordMember.displayName);
     return discordMember;
 }
-
-export async function AsyncGetClanMemberOnlineTime(message, days, discordMention, isDetailed) {
-    var discordName = discordMention == null
-        ? message.member.displayName
-        : GetDiscordMemberByMention(message.guild, discordMention).displayName;
-
-    var apiMember = await AsyncGetMemberByDiscordName(discordName);
-    var clanMember = new ClanMember(apiMember);
-    await clanMember.FetchCharacterIds();
-    clanMember.FetchDiscordMember(message.guild);
-
-    var clanVoiceSummary = await AsyncGetClanVoiceSummary(days);
-    clanMember.AddToVoiceOnline(clanVoiceSummary[clanMember.discordMemberId]);
-
-    var activities = await AsyncGetAllActivities(clanMember, days);
-    activities.forEach(a => clanMember.AddToGameOnline(a.values.timePlayedSeconds.basic.value))
-
-    if (isDetailed) {
-        var detailedVoiceResults = await AsyncGetMemberDetailedVoice(days, clanMember.discordMemberId);
-        var lines = clanMember.FormLinesForDetailedVoice(detailedVoiceResults)
-        message.channel.send(clanMember.GetMemberTimeEmbed(lines));
-    }
-    else message.channel.send(clanMember.GetMemberTimeString());
-}
-
-export async function AsyncGetAllActivities(clanMember, days) {
-    var deltaDate = new Date();
-    deltaDate.setDate(deltaDate.getDate() - days);
-
-    var activities = [];
-    var promises = clanMember.characterIds.map(async characterId => {
-        Array.prototype.push.apply(activities, await AsyncGetCharacterActivities(clanMember, characterId, 0, 'None', deltaDate));
-    });
-    await Promise.all(promises);
-    return activities;
-}
-
-async function AsyncGetCharacterActivities(clanMember, characterId, page, mode, deltaDate) {
-    var filteredActivities = [];
-    var responceActivities = await AsyncGetActivitiesFromApi(clanMember.membershipType, clanMember.membershipId, characterId, page, mode);
-
-    if (responceActivities.ErrorCode == 1665) {
-        clanMember.access = false;
-        return [];
-    }
-    if (!responceActivities.Response?.activities) {
-        return [];
-    }
-
-    var isLastPage = false;
-    responceActivities.Response.activities.forEach(function (activity) {
-        if (deltaDate < new Date(activity.period)) {
-            filteredActivities.push(activity);
-        } else {
-            isLastPage = true;
-        }
-    });
-    if (!isLastPage) Array.prototype.push.apply(filteredActivities, await AsyncGetCharacterActivities(clanMember, characterId, ++page, mode, deltaDate));
-    return filteredActivities;
-}
