+++ commands/restrictedCommand.js
@@ -4,7 +4,7 @@ import config from "../config.json";
 import { CommandManager } from "../commandManager.js"
 import { ShowRecordStatistics, ShowTopTriumphScore } from "../clan/clan.js"
 import { AsyncRoles } from "../clan/clanMember/roles/roles.js"
-import { AsyncGetClanMemberOnlineTime } from "../clan/clanMember/clanMember.js";
+import { AsyncGetClanMemberOnlineTime } from "../clan/clanMember/getOnlineTime.js";
 import { ShowMedalsSummary } from "../discordFeatures/show/showMedalsSummary.js";
 import { ChangeChannelCap } from "../discordFeatures/change/changeChannelCap.js";
 import { ChangeRegion } from "../discordFeatures/change/changeRegion.js";
